chapter_001/src/main/java/ru/job4j/calculate/Converter.java
package ru.job4j.calculate;

/**
 * Корвертор валюты.
 */
public class Converter {
    private double result;

    /**
     * Конвертируем рубли в евро.
     *
     * @param value рубли.
     * @return Евро.
     */
    public int rubleToEuro(int value) {
        return value / 70;
    }

    /**
     * Конвертируем евро в рубли.
     *
     * @param value евро.
     * @return Рубли.
     */
    public int euroToRuble(int value) {
        return value * 70;
    }

    /**
     * Конвертируем рубли в доллары.
     *
     * @param value рубли.
     * @return Доллары
     */
    public int rubleToDollar(int value) {
        return value / 60;
    }

    /**
     * Конвертируем доллары в рубли.
     *
     * @param value доллар.
     * @return Рубли.
     */
    public int dollarToRuble(int value) {
        return value * 60;
    }
}
